requirements.txt
Flask==1.1.2
gunicorn==20.1.0
PyQRCode==1.2.1
vobject==0.9.6.1
google-cloud-storage==2.5.0